requirements.txt
flask==2.0.1
google-cloud-language==2.0.0